Add PCI bus support, device driver support
Most drivers are enabled as modules. Ext2 file system support is hardcoded
 for testing booting/installing on hard disks.
Most of these changes are debatable and customizable.

All built modules are part of the auxiliary floppy, so big busybox
and the kernel modules will share the 1.44MB floppy.

As a reminder: To make the most use of the space on the floppy, it is
using SquashFS to compress the entirety of the FileSystem. All UID/GID are normalized
to 0 (root), as there is no multi-user support in the current kernel.